+++ src/AssemblyInfo.cs
@@ -1,0 +1,8 @@
+using System.Reflection;
+
+[assembly: AssemblyTitle("NapeBar")]
+[assembly: AssemblyProduct("NapeBar")]
+[assembly: AssemblyDescription("Battery monitor for Keychron Nape Pro")]
+[assembly: AssemblyVersion("0.1.7.0")]
+[assembly: AssemblyFileVersion("0.1.7.0")]
+[assembly: AssemblyInformationalVersion("0.1.7")]